fix warnings; add complex_to_real(hoNDImage<T, D>& cplx) function
Former-commit-id: 5c1f968bd0eac8744d2a6a630b1fb8d540d3cb97

+++ toolboxes/core/cpu/arma_math/hoNDImage_util.h
@@ -84,7 +84,7 @@ namespace Gadgetron
 
     /// perform midian filter
     /// w is the window size
-    template<class ArrayType> bool filterMedian(const ArrayType& img, unsigned int w[], ArrayType& img_out);
+    template<class ArrayType> bool filterMedian(const ArrayType& img, size_t w[], ArrayType& img_out);
 
     /// downsample the image by a ratio
     /// new image size = image size / ratio
